Put skill header and progress bar in a container.
The container is called skill__button. This is to easily allow
the clickable area to extend down to the progress bar, opposed
to before when the clickable area was only the header.

The other solution of putting the progress bar inside the header
container will not work because it will interfere with the css
sty;ing for positioning the title and percentage text.